# server/src/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Gachitda API
  version: 1.0.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
security:
  - bearerAuth: []
paths:
...
  /api/noti/subscription:
    post:
      tags:
        - Subscription
      summary: 웹 알림 허용
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subscription:
                  type: object
      responses:
        '200':
          description: 구독 저장 성공
        '401':
          description: INVALID_ACCESS_TOKEN
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                missing_authorization_header:
                  value:
                    error: MISSING_AUTHORIZATION_HEADER
                invalid_authorization_format:
                  value:
                    error: INVALID_AUTHORIZATION_FORMAT
                invalid_access_token:
                  value:
                    error: INVALID_ACCESS_TOKEN
...
  /api/user:
    get:
      tags:
        - User
      summary: 유저의 기본 정보
      responses:
        '200':
          description: 기본 정보
          content:
            application/json:
              schema:
                type: object
                properties:
                  nickname:
                    type: string
                  profileimage:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
                  subscription:
                    type: boolean
        '401':
          description: INVALID_ACCESS_TOKEN
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                missing_authorization_header:
                  value:
                    error: MISSING_AUTHORIZATION_HEADER
                invalid_authorization_format:
                  value:
                    error: INVALID_AUTHORIZATION_FORMAT
                invalid_access_token:
                  value:
                    error: INVALID_ACCESS_TOKEN
...
  /api/user/posts:
    get:
      tags:
        - User
        - Post
      summary: 유저가 쓴/신청한 포스트 리스트
      responses:
        '200':
          description: 포스트 ID & 간단한 정보
          content:
            application/json:
              schema:
                type: object
                properties:
                  writtenPosts:
                    type: array
                    items:
                      type: object
                      properties:
                        postId:
                          type: string
                        category:
                          type: string
                        title:
                          type: string
                        shortDescription:
                          type: string
                        appliedTalents:
                          type: integer
                        address:
                          type: string
                        status:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                        teachAt:
                          type: string
                          format: date-time
                  appliedPosts:
                    type: array
                    items:
                      type: object
                      properties:
                        postId:
                          type: string
                        writerName:
                          type: string
                        category:
                          type: string
                        title:
                          type: string
                        shortDescription:
                          type: string
                        address:
                          type: string
                        status:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                        teachAt:
                          type: string
                          format: date-time
        '401':
          description: INVALID_ACCESS_TOKEN
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                missing_authorization_header:
                  value:
                    error: MISSING_AUTHORIZATION_HEADER
                invalid_authorization_format:
                  value:
                    error: INVALID_AUTHORIZATION_FORMAT
                invalid_access_token:
                  value:
                    error: INVALID_ACCESS_TOKEN
  /api/post:
    post:
      summary: 글 작성
      tags:
        - Post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                category:
                  type: string
                shortDescription:
                  type: string
                detailedDescription:
                  type: string
                address:
                  type: string
                teachAt:
                  type: string
                  format: date-time
      responses:
        '201':
          description: 글 작성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  postId:
                    type: string
                    description: 생성된 글의 ID
        '400':
          description: MISSING_REQUIRED_FIELDS
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                MISSING_REQUIRED_FIELDS:
                  value:
                    error: MISSING_REQUIRED_FIELDS
        '401':
          description: INVALID_ACCESS_TOKEN
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
              examples:
                missing_authorization_header:
                  value:
                    error: MISSING_AUTHORIZATION_HEADER
                invalid_authorization_format:
                  value:
                    error: INVALID_AUTHORIZATION_FORMAT
                invalid_access_token:
                  value:
                    error: INVALID_ACCESS_TOKEN
...